Revert "Redesign layout according to discussion"
