--- antdesigndemo/.babelrc
@@ -1,9 +1,0 @@
-{
-  "presets": ["react-app"],
-  "plugins": [
-    [
-      "import",
-      { "libraryName": "antd", "libraryDirectory": "es", "style": true }
-    ]
-  ]
-}
