--- antdesigndemo/.firebase/hosting.YnVpbGQ.cache
@@ -1,13 +1,0 @@
-asset-manifest.json,1554647537303,c16feca6f1a36f98c64452f08e5db1eabc50f272f73a1a497eb89e11841d1cc7
-manifest.json,1554647509293,77ef04b67a75a750001738b09a30c88fd7b2d6b4081efd92985bacee370b7f49
-index.html,1554647537269,a78bf6a992fddc3eb9584dd8e73f8754b219f307d057df80c24825ea3c9aa86c
-precache-manifest.b2adc94bdf3a2083d072cd4d563d54e9.js,1554647537274,883e6d49030249a7f7f85e10b934fdc6c1579a433cadec0eafd0d0a4f15bdfce
-service-worker.js,1554647537277,8c85240e99b9e1a2064c7d3952f4dc1d2d6973eb50f97b7e0ee8006d3eefd385
-favicon.ico,1554647509290,c599b7a91ab3627e3538125d9f40adc2d4bf949046984262670545dc7738af06
-static/js/runtime~main.fdfcfda2.js,1554647537350,f8959cf2602bcd38f3bd32af487cb68a181f068bd902a6d745e0afa169e5ffd8
-static/js/main.3af2c85b.chunk.js.map,1554647537354,9a9c45c0d91e230cb1dc59c2005a205b0d9658d096494f8aa1997de81ee06921
-static/js/runtime~main.fdfcfda2.js.map,1554647537649,a8a401455e26ed86abc61a092754b78be748d60b7d820f16ae66247076ffb9b6
-static/js/main.3af2c85b.chunk.js,1554647537348,d3b91c547633edc35d13afd6aae0017b017fd873b36c3711cbde881d65d4047d
-color.less,1554647537299,a965907434781884911ee1d09f150cf6d62ad6203378e13d8b15d339a613e46b
-static/js/2.ff517b18.chunk.js,1554647537641,7c8be57838b62312a3d4990db1901372421ca408b305b0888eee6d7427e4b168
-static/js/2.ff517b18.chunk.js.map,1554647537648,ac9dff79bfbf615e3766d9f31282dd03ddbbe8b7778e2564b52135d2fa4929bc
