--- antdesigndemo/.firebaserc
@@ -1,6 +1,0 @@
-    
-{
-  "projects": {
-    "default": "antd-live-theme"
-  }
-}
