+++ antdesigndemo/package.json
@@ -4,33 +4,29 @@
   "private": true,
   "dependencies": {
     "@babel/polyfill": "^7.4.4",
-    "antd": "^3.20.1",
+    "antd": "^3.19.8",
     "antd-theme-webpack-plugin": "^1.3.0",
-    "axios": "^0.19.0",
     "babel-plugin-import": "^1.12.0",
-    "core-js": "^3.1.4",
-    "eslint-config-react-app": "^4.0.1",
+    "core-js": "^3.0.1",
+    "customize-cra": "^0.2.14",
     "history": "^4.9.0",
     "less": "^3.9.0",
     "less-loader": "^5.0.0",
-    "less-vars-to-js": "^1.3.0",
     "prop-types": "^15.7.2",
     "react": "^16.8.6",
-    "react-color": "^2.17.3",
+    "react-app-rewired": "^2.1.3",
     "react-dom": "^16.8.6",
     "react-redux": "^7.1.0",
-    "react-responsive": "^7.0.0",
-    "react-router": "^5.0.1",
     "react-router-dom": "^5.0.1",
     "react-scripts": "3.0.1",
-    "redux": "^4.0.3",
+    "redux": "^4.0.1",
     "redux-thunk": "^2.3.0",
     "regenerator-runtime": "^0.13.2"
   },
   "scripts": {
-    "start": "rescripts start",
-    "build": "rescripts build",
-    "test": "rescripts test",
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
@@ -51,16 +47,8 @@
     ]
   },
   "devDependencies": {
-    "@babel/preset-env": "^7.5.4",
-    "@rescripts/cli": "0.0.11",
-    "@rescripts/rescript-env": "0.0.10",
-    "@rescripts/rescript-use-babel-config": "0.0.8",
-    "firebase-tools": "^7.0.2",
+    "@babel/preset-env": "^7.4.5",
     "http-proxy-middleware": "^0.19.1",
     "redux-logger": "^3.0.6"
-  },
-  "rescripts": [
-    ".rescriptsrc.js"
-    ]
- 
+  }
 }
