+++ antdesigndemo/src/App.js
@@ -1,12 +1,10 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { Router, Route, Switch,Redirect } from 'react-router-dom';
-// import Home from './views/home/index';
-import HomePage from './views/home/index';
+import { Router, Route, Switch } from 'react-router-dom';
+import Home from './views/home/index';
 import { history } from './utils/history';
 import Login from './views/login/Login';
-import { setHtmlFontSize, resetHtmlFontSize } from './utils/setHtmlFontSize';
-// import Nofound from './views/home/Nofound';
+import { setHtmlFontSize, resetHtmlFontSize } from './utils/setHtmlFontSize'
 class App extends Component {
   constructor(props) {
     super(props);
@@ -23,11 +21,12 @@ class App extends Component {
   render() {
     setHtmlFontSize()
     return (<div>
+     
       <Router history={history}>
         <Switch>
-         <Route path={`${process.env.PUBLIC_URL}/login`} component={Login}></Route>
-          <Route  path={`${process.env.PUBLIC_URL}/`} component={HomePage}></Route>
-          {/* <Route  component={Nofound}></Route> */}
+        <Route path={`${process.env.PUBLIC_URL}/login`} component={Login}></Route>
+          <Route exect path={`${process.env.PUBLIC_URL}/`} component={Home}></Route>
+         
         </Switch>
       </Router>
     </div>);
@@ -43,5 +42,5 @@ const mapDispatchToProps = (dispatch) => {
 
   }
 }
-
+// export default App
 export default connect(mapStateToProps, mapDispatchToProps)(App);
