+++ antdesigndemo/src/index.css
@@ -1,162 +1,82 @@
-@charset "utf-8";
-@import './styles/index.less';
-  html{
-    font-size: 16px;
-    height: 100%;
-  }
-  body {
-    margin: 0;
-    background-color: #f0f2f5 !important;
-    height: 100%;
-  }
-  .ant-layout .trigger {
-    font-size: 18px;
-    line-height: 64px;
-    padding: 0 24px;
-    cursor: pointer;
-    transition: color 0.3s;
-  }
-  
-  .ant-layout .trigger:hover {
-    color: #1890ff;
-  }
-  .ant-layout .logo {
-    height: 32px;
-    background: rgba(255, 255, 255, 0.2);
-    margin: 16px;
-  }
-  .steps-content {
-    margin-top: 16px;
-    border: 1px dashed #e9e9e9;
-    border-radius: 6px;
-    background-color: #fafafa;
-    min-height: 100px;
-    padding: 10px 30px 0;
-  }
-  
-  .steps-action {
-    margin-top: 24px;
-  }
-  .section{
-    margin-bottom:20px;
-  }
-  .table-operations {
-    margin-bottom: 16px;
-  }
-  .login{
-    width: 30%;
-    display: inline-block;
-    margin: 10% 35%;
-    font-size: 2rem
-  
-  }
-  .login .login-logo{
-    text-align: center;
-    font-size: 2rem;
-    font-weight: 800;
-    margin-bottom: 20px;
-    color:#fff;
-    font-family: 'Microsoft YaHei';
+@import './styleLess/index.less';
 
-  }
-  .table-operations > button {
-    margin-right: 8px;
-  }
-  .login .login-form {
-    max-width: 14rem;
-    margin: 0 auto;
-  }
-  .login .login-form-forgot {
-    float: right;
-  }
-  .login .login-form-button {
-    width: 100%;
-  }
-  /* fix placeholder is not aligned input in IE */
-  .ant-input{
-    line-height: 1.5
-  }
-  .logo{
-    font-size: 26px;
-    font-family: 700;
-    color: #fff;
-    text-align: center;
-    vertical-align: middle;
-    line-height: 32px;
-  }
-  
-  
-  .stepper .stepper-title{
-    margin-bottom: 0.5rem;
-    display: inline-block;
-    font-family: 800;
-    font-size: 0.9rem
-  }
-  .stepper .stepper-detail>span{
-    display: inline-block;
-    margin-right: 1rem;
-  }
+html{
+  font-size: 16px
+}
 
-  /* theme */
-  .theme-color-content .theme-color-block{
-    float: left;
-    width: 20px;
-    height: 20px;
-    margin-right: 8px;
-    color: #fff;
-    font-weight: 700;
-    text-align: center;
-    border-radius: 2px;
-    cursor: pointer;
-  }
-  .ant-layout-footer{
-    padding: 16px 50px;
-  }
-  .ant-layout-header{
-    padding:'0 20px';
-    
-  }
-  .ant-row .ant-col:nth-child(1) .ant-card{
-    background-color:#015db4;
-    border-color: #187da0;
-  }
-  .ant-row .ant-col:nth-child(2) .ant-card{
-    background-color:#2798ff;
-    border-color: #2eadd3;
-  }
-  .ant-row .ant-col:nth-child(3) .ant-card{
-    background-color:#3dc676;
-    border-color: rgb(124, 220, 164);
-  }
-  .ant-row .ant-col:nth-child(4) .ant-card{
-    background-color:#ffae22;
-    border-color: #f3a318;
-  }
-  .ant-row .ant-col .ant-card:hover{
-    -webkit-transform: translate(0,-15px);
-    -moz-transform: translate(0,-15px);
-    transform: translate(0,-15px);
-    -webkit-box-shadow: 0 5px 10px 0 rgba(0, 0, 0, 0.4);
-    -moz-box-shadow: 0 5px 10px 0 rgba(0, 0, 0, 0.4);
-    box-shadow: 0 5px 10px 0 rgba(0, 0, 0, 0.4);
-  }
-  .ant-row .ant-col:nth-child(1) .ant-card:hover {
-    background: #05529a;
-  }
-  .ant-row .ant-col:nth-child(2) .ant-card:hover {
-    background:#2590f1;
-  }
-  .ant-row .ant-col:nth-child(3) .ant-card:hover {
-    background:#2fbb69;
-  }
-  .ant-row .ant-col:nth-child(4) .ant-card:hover {
-    background:#ef9e14;
-  }
+.ant-layout .trigger {
+  font-size: 18px;
+  line-height: 64px;
+  padding: 0 24px;
+  cursor: pointer;
+  transition: color 0.3s;
+}
 
-  .ant-table-tbody > tr:hover{
-    -webkit-box-shadow: 0 3px 5px 0 rgba(0, 0, 0, 0.4);
-    -moz-box-shadow: 0 3px 5px 0 rgba(0, 0, 0, 0.4);
-    box-shadow: 0 3px 5px 0 rgba(0, 0, 0, 0.4);
-  }
+.ant-layout .trigger:hover {
+  color: #1890ff;
+}
+.ant-layout .logo {
+  height: 32px;
+  background: rgba(255, 255, 255, 0.2);
+  margin: 16px;
+}
+.steps-content {
+  margin-top: 16px;
+  border: 1px dashed #e9e9e9;
+  border-radius: 6px;
+  background-color: #fafafa;
+  min-height: 100px;
+  /* text-align: center; */
+  padding: 10px 30px 0;
+}
+
+.steps-action {
+  margin-top: 24px;
+}
+.section{
+  margin-bottom:20px;
+}
+.table-operations {
+  margin-bottom: 16px;
+}
+.login{
+ width:20%;
+  display: inline-block;
+  margin:10% 40%
+}
+.table-operations > button {
+  margin-right: 8px;
+}
+.login .login-form {
+  max-width: 300px;
+}
+.login .login-form-forgot {
+  float: right;
+}
+.login .login-form-button {
+  width: 100%;
+}
+/* fix placeholder is not aligned input in IE */
+.ant-input{
+  line-height: 1.5
+}
+.logo{
+  font-size: 26px;
+  font-family: 700;
+  color: #fff;
+  text-align: center;
+  vertical-align: middle;
+  line-height: 32px;
+}
 
 
+.stepper .stepper-title{
+  margin-bottom: 0.5rem;
+  display: inline-block;
+  font-family: 800;
+  font-size: 0.9rem
+}
+.stepper .stepper-detail>span{
+  display: inline-block;
+  margin-right: 1rem;
+}
