--- antdesigndemo/src/styles/index.less
@@ -1,36 +1,0 @@
-
-.secondary-color {
-  color: @secondary-color;
-}
-.ant-card{
-  border: 1px solid #c8ced3;
-  height:160px;
-  line-height: 160px;
-  text-align: center;
-  cursor: pointer;
-}
-.ant-card-body{
-  color: #fafafa;
-  font-size: 28px;
-}
-#components-layout-demo-fixed{
-  background: #f0f2f5;
-}
-#components-layout-demo-fixed .logo {
-  width: 120px;
-  height: 31px;
-  background: rgba(255, 255, 255, 0.2);
-  margin: 16px 24px 16px 0;
-  float: left;
-}
-.upload-list-inline .ant-upload-list-item {
-  float: left;
-  width: 200px;
-  margin-right: 8px;
-}
-.upload-list-inline .ant-upload-animate-enter {
-  animation-name: uploadAnimateInlineIn;
-}
-.upload-list-inline .ant-upload-animate-leave {
-  animation-name: uploadAnimateInlineOut;
-}
