--- antdesigndemo/src/styles/variables.less
@@ -1,11 +1,0 @@
-/*@import "~antd/es/style/themes/default.less";*/ 
-@import "~antd/es/style/themes/default.less";
-/*Make sure that all variables initial value must be unique otherwise changing one color will affect the other.*/ 
-@primary-color: #1890ff;
-@secondary-color: #ccc;
-@layout-header-background: #000001;
-@layout-sider-background: #000001;
-@text-color: fade(@black, 65%);
-@heading-color: fade(#000, 85%);
-@text-color-secondary: fade(#1DA57B, 65%);
-@processing-color: fade(@primary-color, 1%);
