--- antdesigndemo/src/views/home/complaints/Complaints.js
@@ -1,15 +1,0 @@
-import React, { Component } from 'react';
-
-class Complaints extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {  }
-    }
-    render() { 
-        return (<div>
-            Complaints
-        </div>  );
-    }
-}
- 
-export default Complaints;
