--- antdesigndemo/src/views/home/management/index.js
@@ -1,14 +1,0 @@
-import React, { Component } from 'react';
-class Management extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {  }
-    }
-    render() { 
-        return ( <div>
-            Management
-        </div> );
-    }
-}
- 
-export default Management;
